Add files via upload
We wanted to see how different variables would affect the cost of housing in Boston. Our dataset comprises a diverse set of input features, each offering unique insights into the socioeconomic landscape of different neighborhoods in Boston. These features include:
1) CRIM: per capita crime rate by town
2) ZN: proportion of residential land zoned for lots over 25,000 sq.ft.
3) INDUS: proportion of non-retail business acres per town
4) CHAS: Charles River dummy variable (1 if tract bounds river; 0 otherwise)
5) NOX: nitric oxides concentration (parts per 10 million) [parts/10M]
6) RM: average number of rooms per dwelling
7) AGE: proportion of owner-occupied units built prior to 1940
8) DIS: weighted distances to five Boston employment centres
9) RAD: index of accessibility to radial highways
10) TAX: full-value property-tax rate per $10,000 [$/10k]
11) PTRATIO: pupil-teacher ratio by town
12) B: The result of the equation B=1000(Bk - 0.63)^2 where Bk is the proportion of blacks by town
13) LSTAT: % lower status of the population
Output variable:
1) MEDV: Median value of owner-occupied homes in $1000's [k$]
AGE: proportion of owner-occupied units built prior to 1940

Using the skills we learned in DS110, we took a deeper look into the data we collected to get an answer on the key factors influencing the median value of owner-occupied homes in Boston, and how accurately we could predict those prices.